requirements.txt
beautifulsoup4==4.13.4
pandas==2.3.0
playwright==1.52.0
requests==2.32.4
selenium-wire==5.1.0
pytest==8.4.1